Add a --shrink flag, and improve Python 3 compatibility of that code
This adds the ability to use the gdishrink function from the command line, and additionally significantly improves the Python 3 compatibility of that code.

First up, it encodes some stuff which was strings as bytes.

Secondly, it fixes one instance where a calculation could sometimes result in a float value, which I believe Python 2 would transparently truncate to an int if needed, but Python 3 requires explicit truncation.

I'm not hugely familiar with getopt, so if there's a better way to insert these functions I'm all ears.